Cleaned up a lot of code for Soilgun
Broke up the Soilgun Soil godclass into seperate classes, all inheriting from BaseSoilProjectile, and cleaned up code all around
